+++ studio/src/schemaTypes/documents/page.ts
@@ -1,6 +1,6 @@
 import {defineField, defineType} from 'sanity'
 import {DocumentIcon} from '@sanity/icons'
-import {pageBuilderSectionMembers} from '../objects/pageBuilderSections'
+import {pageBuilderInsertMenuViews, pageBuilderSectionMembers} from '../objects/pageBuilderSections'
 
 /**
  * Page schema.  Define and edit the fields for the 'page' content type.
@@ -48,14 +48,7 @@ export const page = defineType({
       of: pageBuilderSectionMembers,
       options: {
         insertMenu: {
-          // Configure the "Add Item" menu to display a thumbnail preview of the content type. https://www.sanity.io/docs/studio/array-type#efb1fe03459d
-          views: [
-            {
-              name: 'grid',
-              previewImageUrl: (schemaTypeName) =>
-                `/static/page-builder-thumbnails/${schemaTypeName}.webp`,
-            },
-          ],
+          views: pageBuilderInsertMenuViews,
         },
       },
     }),